Add _initialize_make function
Exiting mid-script from microservices in this repo didn't previously allow the system to perform cleanup. (For example, the operator might get "stuck" with white or red text in their terminal window.) This function helps the system gracefully exit the script. It will only run if the operator presses ctrl+C.