No mkdir till derivative creation starts (makeDer)
Moves the step of creating the output folder to AFTER the operator interacts with the prompts -> this means that if the operator ctrl+C's the process before creating the derivative, there won't be an empty access folder in the package. This will provide another hint for anyone visually inspecting the package for the presence of derivatives.