1-reel syntax + variables in restructureDPX
Made changes to the section of restructureDPX that deals with DPX packages of one-reel works.
-Replaced redundant find/exec one-liners with simple mv commands (the find/exec commands returned "No such file or directory" errors).
-Replaced short variables (mov, wav, dpx) with longer ones (movpath, wavpath, dpxpath) to give more clarity about their usage.

Note that the script (both before and after this commit) can't handle one-reel DPX packages with multiple directories named *WAV, *MOV, *DPX—it will only find and execute commands on the first directory it finds of each type.