restructureDPX: create non-interview option
restructureDPX is a script that currently relies on interviewing the operator and asking for input (source, destination, ID #, etc.). The operator now has the option to pass that input as flags instead, which also allows restructureDPX to be run with less interactivity (e.g. it can now be called by other scripts).