ingestfile: add profile for Great Migration
Current process: 1. rsyncs package as-is to $destination/$mediaid/$objects 2. generates metadata files for video files not in "ACCESS" subfolders 3. does not run makeH264 because expects that there are already derivatives in "ACCESS" subfolders.

Result is AIP with "objects" folder (holding original SIP) and "metadata" folder (holding files created in makemetadata).